JPA/SpringBoot와 JPA 활용/jpashop/src/test/java/jpabook/module/member/MemberServiceTest.java
package jpabook.module.member;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;

import javax.sql.DataSource;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.assertThrows;

@Transactional
@SpringBootTest
class MemberServiceTest {

    @Autowired
    MemberService memberService;

    @Autowired
    DataSource dataSource;

    @DisplayName("회원 가입")
    @Test
    public void join () throws Exception {
        // given
        Member member = new Member();
        member.setName("kim");

        // when
        Long saveId = memberService.join(member);

        // then
        assertThat(memberService.findById(saveId)).isEqualTo(member);
    }

    @DisplayName("중복 회원 가입 예외")
    @Test
    public void join_duplicate() throws Exception {
        // given
        Member member1 = new Member();
        member1.setName("kim");

        Member member2 = new Member();
        member2.setName("kim");

        // when
        memberService.join(member1);

        // then
        assertThrows(IllegalStateException.class,()-> memberService.join(member2));
    }

}